add include specs
This patch introduces a notion of include specs to allow
for finer grained control over what data is loaded.

The `pggen` root module grows a new submodule `include` in
this patch which defines the data structure and DSL syntax
for include specs.

The generated code now produces two public methods instead
of three, with the most flexible being `<Entity>FillIncludes`

The documentation included in this patch explains the concept
further.

Closes #8