drop public dep on willf/bitset
Having a third party library as part of your public API
is generally a really bad idea for a lot of reasons, so
this patch shoves willf/bitset into opendoor-labs/pggen
and exposes a new type `pggen.FieldSet` with a much reduced
interface. It is just a thin wrapper around a willf/bitset,
but this should give us the option to evolve the implementation
and protect against upstream breaking changes.