add support for funny names
SQL allows identifiers to be pretty strange through
the use of quoted identifiers. There were previously
a fair number of assumptions baked into the pggen codebase
about sql identifiers being fairly sensible, but we are
dealing with a COBOL flavored language so that isn't
a reasonable assumption.

This patch fixes all those assumptions.

Closes #3