standardize include filling interface
This patch updates the interface for filling in decendant
objects to just consist of the <Entity>FillIncludes and
<Entity>BulkFillIncludes methods. It seems like it was overkill
to have a special method just for filling in all the includes
when we already give people an include spec to do just that.

We also get rid of the single pointer vs list of pointers weirdness
in this patch.

Closes #10