add support for passing and returning arrays
This patch teaches pggen to generate code for queries that
take arrays as parameters and that return arrays in their
result set.

In adding array support, I uncovered a bug in how arrays of
nullable values are deserialized when those nullable values
are represented with a simple box. Because there are no users
yet, I can still afford to swap out the interface to use
`sql.Null<Type>` values for possibly nullable values now.