fix 'SELECT *' bug in gen'd Scan methods
The Problem
===========

Previously, our approach to scanning tables was pretty straightforward.
We would do a `SELECT * FROM table ...` and then pass each of the
table's generated struct fields as parameters to the `sql.Rows.Scan`
method in the order in which the appear in the database _at codgen
time_. @JoseGMorgan recently pointed out to me that this is broken
because the migrations for a production or staging database might
be applied in non-timestamp order.

Suppose you have two migrations:

mirgation_1.sql
```sql
ALTER TABLE foo ADD COLUMN blah text;
```

migration_2.sql
```sql
ALTER TABLE foo ADD COLUMN blip int;
```

Even though migration_1 was created first chronologically, migration_2
gets merged and applied first. Now the production database will have the
table:

```sql
CREATE TABLE foo (
  ...
  blip int,
  blah text
);
```

but if we apply all the migrations in timestamp order we will get

```sql
CREATE TABLE foo (
  ...
  blah text,
  blip int
);
```

the situation is even worse if the columns are of the same type because
then we won't even get a type error to tell us that something funky is
going on.

The Fix
=======

This patch resolves the issue by adding some tables to the `PGClient`
that we can use to reconcile the difference between the codegen time
and runtime positions of the different columns on a table. Now, the
first time a record is scanned into a particular table struct we
fill in the relevant index table and use it to make sure that we
load the right stuff.

The Breakage
============

This patch is a breaking change because we switch to no longer exposing
the database handle field for `PGClient`. Previously, `PGClient` was
just a very lightweight wrapper around the underlying database
connection, but now we need to store some auxiliary data, so we can't
let users just manipulate the database connection field directly
anymore.

We also need a real database connection, not a transaction so you
can no longer create a PGClient out a `sql.Tx`.

The field no longer being public and not being allowed to be a `sql.Tx`
sort of demands a new API for working with database transactions, so
`PGClient` grew one of those in this patch as well.

Closes #40