add Upsert method
This patch adds new Upsert and BulkUpsert methods to the generated
PGClient. The upsert methods allow users to determine the conflict
targets and the fields which will get updated in the event of a
conflict. This doesn't expose the full power of postgres upsert,
but it should cover the most common case.

Closes #44