fix: add short-circuit to bulk methods
This patch adds some short circuit conditionals to all the
bulk endpoints which just check if we have been passed an
empty list and immediatly return. This makes dynamic SQL generation
simpler because we don't have to special case empty lists.
In most cases, I think this is just an optimization, but it fixes
a real bug for upsert.