fix bug in the way that cyclic data is loaded
This patch adds some tests for cyclic object graphs and
fixes everything needed to make them pass. There are a bunch
of different fixes required in order to make this work. Some
highlights are:

  - We teach include specs to handle cycles while printing themselves.
  - Include specs how have a documented way to represent cycles.
  - We thread all of the rows ever loaded through our recursive include
    filling routines rather than just carrying the rows from the most
    recent generation over. This allows us to notice when a row has
    already been loaded and use the object already in memory rather than
    creating a new object with identical data but a different memory
    location.

After this patch the go object graph that pggen gen loads during include
filling faithfully represents the graph of database objects.

This is a breaking change because 1-* relationships were previously
represented with slices of model structs rather than slices of pointers
to model structs, which makes it impossible to faithfully represent
the object graph found in the database.

Closes #51