test against all supported postgres versions in ci
This patch teaches our CI pipeline to test against all the
major supported postgres versions as of today according to
the official postgres website. The postgres system catalog
tables are actually very stable across different postgres
versions, but they are documented as unstable interfaces,
so we need to test against all the versions of postgres that
we want to support.

Closes #14