add support for custom names in `belongs_to`
This patch teaches `pggen` to accept a new configuration option
when declaring a relationship between to tables. The new
`parent_field_name` option controls the name of the field which
is added to the generated model struct which corrisponds to the
table that is refered to by the foreign key in the database.

Closes #59