make references bidirectional
This patch makes references generated by pggen bidirectional both
for more convenient usage and to more accurately reflect the underlying
relational semantics.

`belongs_to` relationships now produce fields in both the parent and
child structs (previously, they would just produce fields in the parent
structs). The include spec machinery has been updated to be able to
correctly fill in the new fields, and the generated *AllIncludeSpec
values now fill in parent pointers as well as child pointers.

Parent pointers are always just that, pointers, not slices (as
references to children can sometimes be). This is because the foreign
key lives in the child table, so it can reference at most one parent.

I extended one of the cycle tests to make sure that works.

Closes #43