add mergeTags routine
This patch introduces a new routine for intelligently
merging two different golang tags. The main thing this
gets us is better gorm compatibility because now user
provided 'gorm' tags will actually be picked up.

For now, the new routine is not hooked up to anything,
but the core algorithm is there.