generate DBQueries interface
This patch teaches pggen to generate a new DBQueries interface
that covers all the methods common to both PGClient and TxPGClient.

This currently duplicates some code and hits the database more than
is entirely optimal.

It also needs a test. I think the way to go is to write a
reflection-based test to make sure that the names and types of
all the methods match up properly.