breaking: unify upsert/insert API for pkeys
Hopefully this will be the last breaking change for a while.
The main breakage is that Upsert now uses the same style of
"UsePkey" functional option for handling the insertion of
primary keys rather than relying on the presence of a flag
in the provided bitset. I think this is more explicit and
it is certainly more consistant.

I've also taken the opportunity to put in a few other breaking
changes to rip the bandaid off and make it easier to evolve the
API in the future.
  - All generated methods now accept a list of functional options.
    Mostly they don't do anything with these options, but this will
    provide us an extension point in the future.
  - The internal impl struct no longer exports its methods. The struct
    was private already so I'm not sure this is even breaking, but
    maybe there was a way to get at those methods via reflection. This
    just battens down the hatches.

Fixes #70