add first class support for soft deletes
This patch adds first class support for soft deletes in the same
way that we already had support for auto-updating created_at/updated_at
timestamp fields.

@NickPPC and @martinxsliu have both brought up this feature recently,
so I thought it would be a good thing to add.

Previously, in order to support soft-deletes one would have to write
custom queries that include checks on the delete field. This is a little
annoying, but still probably ok as long as you know about it. The truely
nasty part of the previous situation is that you would have had to give
up using the include machinery entirely and re-implment your own, which
is sad.

To get a little more concrete about the semantic changes that this patch
includes:
    - There is now a new `deleted_at_field` config key both in the
      global scope and on table configs. Tables override the global
      scope. The value of this key is expected to be the name of a
      nullable timestamp field.
    - The get and list generated methods now respect the configured
      soft-delete timestamp.
    - The delete generated method now soft-deletes rather than blowing
      away all the data if a soft-delete has been configured. The
      DeleteDoHardDelete option from the `pggen` pacakge overrides this
      behavior.
    - The include machinery respects the soft-delete timestamp.

Closes #97