pin msgv (minimum supported go version) in ci (#102)
This patch establishes an minimum supported go version of
1.11 by pinning it in CI. Pinning a minimum language version
is mostly a thing from the rust community because of how
aggressively rustc releases, but I think it is a good idea
in general. For a library, using a new function when you have
users on an old language version can be seen as a breaking change,
though it is a bit of a grey area. pggen will not commit to never
increasing the msgv, but it won't do so without a good reason and it
will never do so accidentally (it should be at least mentioned in
release notes or a git commit message).

In order to fix the lint CI step, we bump golangci-lint in this patch
as well.

This patch also adds a note to HACKING.md about some of `pggen`s design
philosophy.

Closes #33