run examples as tests
This patch turns the collection of examples into runnable tests.
Now the examples will be run when you do:

```
go generate ./... && go test -p 1 ./...
```

An example test checks that the generated code looks like what is
checked in in addition to making sure that the output looks correct.

Because example tests care about the output of the code generated,
I finally tracked down why generated code was not stable across
different machines (#53). The problem was in the way we were printing
the type set (ordering was based on a hash).

Closes #106
Fixes #53