update middleware
This patch updates the new middleware stuff to rip out the
mocks. The new tests are shorter.

It also changes the runtime behavior and semantics of the connection
wrapper a little. Previously, we would call the middleware factory
methods on *every invocation* of the primitive database access methods.
This is not super obvious from a users perspective and a potential
performace footgun (we should not be constructing a new closure each
time we execute a query against the database). Now, the factory method
is called eager when the middleware is registered.

Additionaly, we now make it so that middlewares stack, so you can
repeatedly call WithExecMiddleware to build up a chain of middleware.
The first time around there was some exterior utilities for composing
the middleware that I shied away from because of the extra API surface,
but this does not really change the API surface and gives us a more
conveniant API.