fix race in lazy computation of column offset tab
This patch fixes a race in the way that the column offset
tables were getting computed. Previously, multiple goroutines
could attempt to Scan a record of the same type at once and
then compute the offset table at the same time. They could
then write different slice headers at the same time, potentially
corrupting data.

The fix is to guard each offset table with a RWMutex. The scan
routine acquires a read lock whenever it accesses the table and
the update routine acquires a write lock before updating it.
We could make the critical section for the update routine shorter
(by just wrapping the actual slice update), but I thought it would
be better to prevent readers from enqueueing more writers and
causing thrash. I did not really benchmark this.

This race is unlikely to occur unless a service is under heavy
load, but I was able to write a test that confirms that it is
in fact a real issue.

Threading code is tricky, so when we audit this for bugs, we want
to make sure that:
    - Scan has released all locks by the time it returns (I needed
      fairly fine grained control of when the lock was held so I
      had to kick it C style rather than using defer).
    - No read lock can ever be held by a given goroutine when it calls
      fillColPosTab, as this would deadlock.

Fixes #115