add support for using jackc/pgx as a driver (#122)
This patch updates `pggen` to support using `jackc/pgx` as
a database driver. `lib/pq` has entered maintenance mode, and
`jackc/pgx` is the recommended replacement. I still cannot
recommend using `jackc/pgx` as a driver at this time because it
has [a bug][1] that prevents `pggen` from gracefully handling
schema changes.

Most of the changes involve tweaking our scanners to be able
to consume both the data that `lib/pq` and `jackc/pgx` pass
down.

For the moment, the generated code still depends on `lib/pq`
because we need to be able to use `lib/pq.Array` to scan arrays
of values. Once `jackc/pgx` becomes the recommended driver, we
will switch to using whatever the equivalent of `lib/pq.Array`
is for `jackc/pgx` (or roll our own). We will not drop support
for using `lib/pq` as a driver (likely not ever, but certainly
not without a major version bump).

This bumps our go versions, including the msgv. We need to bump
the msgv because jackc/pgx does not support go 1.11.

Closes #119

[1]: https://github.com/jackc/pgx/issues/841