consolidate scan struct implementation
This patch consolidates the code for generating
model structs and query return value structs. This
is nice from a maintanence perspective, but it
also brings a few bugfixes that were only in the
model struct version over to the query version.

In particular, queries of the form `SELECT * FROM ...`
were still vulnerable to schema changes shifting
causing trouble due to gentime vs runtime differences
or migrations run while a service was running.

Closes #121