add support for using defaults in inserts/upserts (#132)
This patch teaches the generated insert and upsert methods
about a new functional configuration option for both the insert
and upsert methods. This option allows you to specify a set of
fields which should have their values pulled from the defaults
specified in the database table definition. Hopefully this will
encourage people to lean on the database for defaults a little
more.