teach pggen to generate shims for json types
This patch teaches pggen to automatically convert between
arbitrary go types and `json` or `jsonb` types in postgres.
To do this we use the stdlib `encoding/json` package, so
all the standard techniques for working with json in go
can be applied here.

The patch comes with an example showing how to do it.

Closes #67