validate import strings
Previously specifying import strings for type overrides and
json types was a little bit fraught because if you did not
get the format right it could cause an internal error when
we went to gofmt the code before landing it on disk.

This patch fixes that by adding a validation step right
after the config file is read in.

Closes #134