rip out prepared function support
This patch removes support for prepared functions. I've found that
in practice, we never actually use this feature, as the `query`
feature tends to be easier to use (managing queries via migrations
is a pain it turns out). I want to cut down the public interface
of pggen before we publicly release it to reduce the maintanence
burden and present a simpler tool.

I'm a little sad about this because pggen started its life as nothing
but a tool to read prepared functions and turn them into query shims.
In fact, that last time I looked at ripping out prepared function
support I decided not to because query support was still implemented
in terms of lots of the prepared function machinery.

Closes #153