switch examples to use jackc/pgx
This is step one in the migration away from lib/pq over to
jackc/pgx. We still need to switch the library that the tool
itself uses and prevent the generated code from depending on
lib/pq (or any other driver library). I'm thinking I should
probably fork out the scanner types into a seperate library
and put it in ethanpailes/pgtypes or something. That way people
can depend on just the scanner types that they want without
having to depend on a deprecated library.

Makes progress on #155