switch the pggen tool to jackc/pgx
As part of the continued campaign to migrate from lib/pq
to jackc/pgx, this patch switches the database driver that
pggen itself uses when it is connecting to the database to
get at the meta tables. It seems that the two divers translate
one of the more esoteric system types differently, so some
code had to be adapted for that, but otherwise there is no
need to adapt.