remove default dep on satori/go.uuid
This patch removes the default use of `satori/go.uuid` as the
uuid library. In the process, we fix a bug with how null
conversions for type overrides were handled.

This patch adds a new example since using a UUID type is a
common desire and I bet users will want some guidance on how
to do it.

Closes #133