IsNotFoundError: check causal chain as well
This is a breaking change, but it shouldn't impact any
current existing usage. This will be nice because it means
you can intercept not-found errors in middleware and wrap them
with your own error tagging system and calls to IsNotFoundError
in your existing codebase should continue to work.

This also makes us match the behavior of errors.Is from the standard
library better, which ought to be less confusing given the name
of the function.