middleware: add support for error converter func
This patch adds support for injecting a routine for
converting pggen errors to errors of a different sort.
It is already possible to intercept errors systemically
at the level of interactions with the database driver,
but pggen itself creates a few errors that cannot be
intercepted in this way (including the NotFoundError,
which is probably the most interesting sort of error
that pggen returns).